% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-utils.R
\name{alr}
\alias{alr}
\title{Compute the ALR of a matrix}
\usage{
alr(x, d = NULL)
}
\arguments{
\item{x}{A matrix where the rows are the samples}

\item{d}{Index of column used as a reference. Defaults to last column}
}
\value{
matrix
}
\description{
Compute the ALR of a matrix
}
