% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-utils.R
\name{alr_array}
\alias{alr_array}
\title{Compute the ALR of an array}
\usage{
alr_array(x, d = dim(x)[parts], parts)
}
\arguments{
\item{x}{multidimensional array in simplex}

\item{d}{Index of column used as a reference. Defaults to last column}

\item{parts}{index of dimension of `x` that represents parts}
}
\value{
array
}
\description{
Compute the ALR of an array
}
