% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-utils.R
\name{alrInv}
\alias{alrInv}
\title{Compute the inverse ALR of a matrix}
\usage{
alrInv(y, d = NULL)
}
\arguments{
\item{y}{An ALR transformed matrix}

\item{d}{Index of column used as a reference. Defaults to last column}
}
\value{
matrix
}
\description{
Compute the inverse ALR of a matrix
}
