% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-utils.R
\name{alrInv_array}
\alias{alrInv_array}
\title{Compute the ALR of an array}
\usage{
alrInv_array(y, d = dim(y)[coords] + 1, coords)
}
\arguments{
\item{y}{multidimensional ALR transformed array}

\item{d}{Index of column used as a reference. Defaults to last column}

\item{coords}{index of dimension of `x` that represents coordinates}
}
\value{
array
}
\description{
Compute the ALR of an array
}
