% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidofit_methods.R
\name{as.list.orthusfit}
\alias{as.list.orthusfit}
\title{Convert object of class orthusfit to a list}
\usage{
\method{as.list}{orthusfit}(x, ...)
}
\arguments{
\item{x}{an object of class orthusfit}

\item{...}{currently unused}
}
\value{
A list of the converted orthusfit object
}
\description{
Convert object of class orthusfit to a list
}
