% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidofit_methods.R
\name{as.list}
\alias{as.list}
\title{Generic method to convert to list}
\usage{
as.list(x, ...)
}
\arguments{
\item{x}{An object of class pibblefit or orthusfit}

\item{...}{Other objects to pass}
}
\value{
A list object
}
\description{
Generic method to convert to list
}
