% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-utils.R
\name{clr_array}
\alias{clr_array}
\title{Compute the CLR of an array}
\usage{
clr_array(x, parts)
}
\arguments{
\item{x}{multidimensional array in index}

\item{parts}{index of dimension of `x` that represents parts}
}
\value{
array
}
\description{
Compute the CLR of an array
}
