% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-utils.R
\name{create_default_ilr_base}
\alias{create_default_ilr_base}
\title{Create a default ILR base}
\usage{
create_default_ilr_base(D)
}
\arguments{
\item{D}{the number of parts (e.g., number of columns in untransformed data)}
}
\value{
A matrix
}
\description{
Create a default ILR base
}
