% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fido.R
\docType{package}
\name{fido_package}
\alias{fido}
\alias{fido-package}
\alias{fido_package}
\title{fido: Fitting and Analysis of Multinomial Logistic Normal  Models}
\description{
Provides methods for fitting and inspection of Bayesian Multinomial 
 Logistic Normal Models using MAP estimation and Laplace Approximation.
 Key functionality is implemented in C++ for scalability.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jsilve24.github.io/fido/}
  \item Report bugs at \url{https://github.com/jsilve24/fido/issues}
}

}
\author{
\strong{Maintainer}: Justin Silverman \email{JustinSilverman@psu.edu}

Other contributors:
\itemize{
  \item Kim Roche \email{kimberly.roche@duke.edu} [contributor]
  \item Michelle Nixon \email{pistner@psu.edu} [contributor]
}

}
