% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-utils.R
\name{gather_array}
\alias{gather_array}
\title{Gather Multidimensional Array to Tidy Tibble}
\usage{
gather_array(a, value, ..., .id = NULL)
}
\arguments{
\item{a}{multidimensional array}

\item{value}{unquoted name of column with values (defaults to "var")}

\item{...}{unquoted dimension names (defaults to "dim_1", "dim_2", etc...)}

\item{.id}{if specified, name for column created with name of a captured}
}
\value{
data.frame
}
\description{
Gather Multidimensional Array to Tidy Tibble
}
\examples{
a <- array(1:100, dim =c(10, 5, 2))
gather_array(a, sequence, A, B, C)
}
\seealso{
spread_array
}
