% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallard-data.R
\docType{data}
\name{mallard}
\alias{mallard}
\title{Data from Silverman et al. (2018) Microbiome}
\format{
A list containing "otu_table", "sample_data", "tax_table", and "refseq".
}
\usage{
data(mallard)
}
\description{
High Resolution (hourly and daily) sampling of 4 in vitro artificial gut models
with many technical replicates to identify technical variation.
}
\details{
This data is at the sequence variant level. Data at the family level
processed as in Silverman et al. 2018 is given in \code{\link{mallard_family}}
}
\references{
Silverman et al. "Dynamic linear models guide design and 
  analysis of microbiota studies within artificial human guts". 
  Microbiome 2018 6:202
}
