% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-utils.R
\name{miniclo}
\alias{miniclo}
\title{Closure operator}
\usage{
miniclo(x)
}
\arguments{
\item{x}{vector or matrix (rows are samples, parts are columns) of data in simplex}
}
\value{
x with row entries divided by sum of row (converts vectors to row matricies)
}
\description{
Closure operator
}
\examples{
x <- matrix(runif(30), 10, 3)
x <- miniclo(x)
}
