% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-utils.R
\name{miniclo_array}
\alias{miniclo_array}
\title{Closure Operation applied to array on margin}
\usage{
miniclo_array(x, parts)
}
\arguments{
\item{x}{multidimensional array}

\item{parts}{index of dimension of \code{x} that represents parts (e.g., compositional variables)}
}
\value{
array
}
\description{
Array version of \code{\link{miniclo}}.
}
\examples{
x <- array(1:100, dim=c(10, 5, 2))
miniclo_array(x, parts=2)
}
