% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pibble_plotting.R
\name{plot}
\alias{plot}
\title{Plot Summaries of Posterior Distribution of pibblefit or orthusfit Parameters}
\usage{
plot(x, ...)
}
\arguments{
\item{x}{an object of class pibblefit or orthusfit}

\item{...}{other arguments passed to plot.pibblefit (see details)}
}
\value{
ggplot object
}
\description{
Plot Summaries of Posterior Distribution of pibblefit or orthusfit Parameters
}
