% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidofit_methods.R, R/generics_s3.R
\name{ppc_summary.pibblefit}
\alias{ppc_summary.pibblefit}
\alias{ppc_summary}
\title{Generic Method to Plot Posterior Predictive Summaries}
\usage{
\method{ppc_summary}{pibblefit}(m, from_scratch = FALSE, ...)

ppc_summary(m, ...)
}
\arguments{
\item{m}{model object}

\item{from_scratch}{should predictions of Y come from fitted Eta or from 
predictions of Eta from posterior of Lambda? (default: false)}

\item{...}{other arguments to pass}
}
\value{
vector
}
\description{
Generic Method to Plot Posterior Predictive Summaries
}
