% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidofit_methods.R
\name{predict}
\alias{predict}
\title{Predict response from new data}
\usage{
predict(object, ...)
}
\arguments{
\item{object}{An object of class pibblefit}

\item{...}{Other objects to be passed to the `predict` function}
}
\value{
(if summary==FALSE) array D x N x iter; (if summary==TRUE) 
tibble with calculated posterior summaries
}
\description{
Predict response from new data
}
\examples{
sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
}
