% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidofit_methods.R
\name{print}
\alias{print}
\title{Print dimensions and coordinate system information for an orthusfit or pibblefit object.}
\usage{
print(x, ...)
}
\arguments{
\item{x}{an object of class pibblefit or orthusfit}

\item{...}{other arguments to pass to summary function}
}
\description{
Print dimensions and coordinate system information for an orthusfit or pibblefit object.
}
\examples{
\dontrun{
fit <- pibble(Y, X)
print(fit)
}
}
