% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISK_CCFA.R
\docType{data}
\name{RISK_CCFA_tax}
\alias{RISK_CCFA_tax}
\title{Data from Gevers et al. (2014)}
\format{
A taxonomy table.
}
\usage{
data(RISK_CCFA)
}
\description{
OTU data and metadata for 1,359 samples in a Crohn's disease study
}
\details{
Study is described here: \url{https://pubmed.ncbi.nlm.nih.gov/24629344/}.
Data was obtained from \url{https://github.com/twbattaglia/MicrobeDS}.
}
\references{
Gevers D, et al. The treatment-naive microbiome in new-onset Crohn's disease. Cell Host Microbe. 2014 Mar 12;15(3):382-392. doi: 10.1016/j.chom.2014.02.005. PMID: 24629344; PMCID: PMC4059512.
}
