% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidofit_methods.R
\name{summary.orthusfit}
\alias{summary.orthusfit}
\title{Summarise orthusfit object and print posterior quantiles}
\usage{
\method{summary}{orthusfit}(
  object,
  pars = NULL,
  use_names = TRUE,
  as_factor = FALSE,
  gather_prob = FALSE,
  ...
)
}
\arguments{
\item{object}{an object of class orthusfit}

\item{pars}{character vector (default: c("Eta", "Lambda", "Sigma"))}

\item{use_names}{should summary replace dimension indices with orthusfit 
names if names Y and X were named in call to \code{\link{orthus}}}

\item{as_factor}{if use_names and as_factor then returns names as factors 
(useful for maintaining orderings when plotting)}

\item{gather_prob}{if TRUE then prints quantiles in long format rather than 
wide (useful for some plotting functions)}

\item{...}{other expressions to pass to summarise (using name 'val' unquoted is 
probably what you want)}
}
\value{
A list
}
\description{
Default calculates median, mean, 50\% and 95\% credible interval
}
