% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidofit_methods.R
\name{summary}
\alias{summary}
\title{Summarise pibblefit or orthusfit object and print posterior quantiles}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{an object of class pibblefit or orthusfit}

\item{...}{other objects to be passed to `summary.pibblefit` or `summary.orthusfit`}
}
\value{
A list if class is `pibblefit` or `orthusfit`
}
\description{
Default calculates median, mean, 50\% and 95\% credible interval
}
