% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrbias_mock-data.R
\docType{data}
\name{Y}
\alias{Y}
\title{Data from Silverman et al. (2019) bioRxiv}
\format{
an matrix Y (counts for each community member)
}
\description{
Mock communities and calibration samples created for measuring and validating model
of PCR bias. This data has been preprocessed as in the original manuscript.
}
\references{
Justin D. Silverman, Rachael J. Bloom, Sharon Jiang, Heather K. Durand, Sayan Mukherjee, 
Lawrence A. David. (2019) Measuring and Mitigating PCR Bias in Microbiome Data. 
bioRxiv 604025; doi: https://doi.org/10.1101/604025
}
