% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidofit_methods.R
\name{as.list.pibblefit}
\alias{as.list.pibblefit}
\title{Convert object of class pibblefit to a list}
\usage{
\method{as.list}{pibblefit}(x, ...)
}
\arguments{
\item{x}{an object of class pibblefit}

\item{...}{currently unused}
}
\value{
A list from the converted pibblefit object.
}
\description{
Convert object of class pibblefit to a list
}
