% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidofit_methods.R
\name{coef}
\alias{coef}
\title{Return regression coefficients of pibblefit or orthusfit object}
\usage{
coef(object, ...)
}
\arguments{
\item{object}{an object of class pibblefit or orthusfit}

\item{...}{other options passed to coef.pibblefit  or coef.orthusfit (see details)}
}
\value{
Array of dimension (D-1) x Q x iter
}
\description{
Return regression coefficients of pibblefit or orthusfit object
}
\examples{
sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
coef(fit)

}
