% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fido.R
\docType{package}
\name{fido_package}
\alias{fido_package}
\title{fido: Fitting and Analysis of Multinomial Logistic Normal  Models}
\description{
Provides methods for fitting and inspection of Bayesian Multinomial 
 Logistic Normal Models using MAP estimation 
 (with the ADAM optimizer) and Laplace Approximation. Key functionality is 
 implemented in C++ for scalability.
}
