% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{r2}
\alias{r2}
\alias{r2.pibblefit}
\alias{r2.bassetfit}
\title{Generic Method to Calculate R2 for Fitted Model}
\usage{
r2(m, ...)

\method{r2}{pibblefit}(m, ...)

\method{r2}{bassetfit}(m, ...)
}
\arguments{
\item{m}{model object}

\item{...}{other arguments to pass}
}
\value{
vector
}
\description{
Generic Method to Calculate R2 for Fitted Model
}
\details{
Calculates Posterior over Linear Model R2 as:
  \deqn{1-\frac{SS_{res}}{SS_tot}}
  where \eqn{SS} is defined in terms of trace of variances

Calculates Posterior over Basset Model R2 as:
  \deqn{1-\frac{SS_{res}}{SS_tot}}
}
